# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/v3/accounts/{address}/payments": {
            "get": {
                "description": "根据用户地址查询Payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-query"
                ],
                "summary": "根据用户地址查询Payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账户地址",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "币种",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "币种合约地址",
                        "name": "contract",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始时间",
                        "name": "begin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游标",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序(ASC/DESC)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.V3Payment"
                            }
                        }
                    }
                }
            }
        },
...
        "/v3/ledgers/{height}/payments": {
            "get": {
                "description": "根据区块高度查询Payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-query"
                ],
                "summary": "根据区块高度查询Payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "区块高度",
                        "name": "height",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "币种",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "币种合约地址",
                        "name": "contract",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始时间",
                        "name": "begin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游标",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序(ASC/DESC)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.V3Payment"
                            }
                        }
                    }
                }
            }
        },
...
        "/v3/nodes/{address}": {
            "get": {
                "description": "查询节点信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-dpos"
                ],
                "summary": "查询节点信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "地址",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/types.BeanValidator"
                        }
                    }
                }
            }
        },
        "/v3/nodes/{address}/voters": {
            "get": {
                "description": "查询节点投票人信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-dpos"
                ],
                "summary": "查询节点投票人信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "地址",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.BeanVots"
                            }
                        }
                    }
                }
            }
        },
        "/v3/payments": {
            "get": {
                "description": "查询所有Payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-query"
                ],
                "summary": "查询所有Payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "币种",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "币种合约地址",
                        "name": "contract",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始时间",
                        "name": "begin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游标",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序(ASC/DESC)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.V3Payment"
                            }
                        }
                    }
                }
            }
        },
...
        "/v3/transactions/{txhash}/payments": {
            "get": {
                "description": "根据txhash查询Payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-query"
                ],
                "summary": "根据txhash查询Payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "交易hash",
                        "name": "txhash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "币种",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "币种合约地址",
                        "name": "contract",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始时间",
                        "name": "begin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游标",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序(ASC/DESC)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.V3Payment"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "database.V3Payment": {
            "type": "object",
            "properties": {
                "contract": {
                    "description": "合约地址，原生币为空或全零黑洞地址",
                    "type": "string"
                },
                "createdAt": {
                    "description": "区块时间",
                    "type": "string"
                },
                "evName": {
                    "description": "事件名称",
                    "type": "string"
                },
                "hash": {
                    "description": "交易hash",
                    "type": "string"
                },
                "height": {
                    "description": "区块高度",
                    "type": "integer"
                },
                "id": {
                    "description": "数据库自增id",
                    "type": "integer"
                },
                "idx": {
                    "description": "交易索引",
                    "type": "integer"
                },
                "receiver": {
                    "description": "转账接受方地址",
                    "type": "string"
                },
                "sender": {
                    "description": "转账发起方地址",
                    "type": "string"
                },
                "symbol": {
                    "description": "币种，原生币为“OLO”",
                    "type": "string"
                },
                "value": {
                    "description": "交易金额",
                    "type": "string"
                }
            }
        },
...
        "types.BeanValidator": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "节点公钥",
                    "type": "string"
                },
                "balance": {
                    "description": "余额（累积收益）",
                    "type": "string"
                },
                "genesis": {
                    "description": "是否是创世节点",
                    "type": "boolean"
                },
                "mortgaged": {
                    "description": "已抵押",
                    "type": "string"
                },
                "nonce": {
                    "description": "nonce",
                    "type": "integer"
                },
                "power": {
                    "description": "投票权",
                    "type": "integer"
                },
                "redeem": {
                    "description": "待赎回",
                    "type": "boolean"
                },
                "runFor": {
                    "description": "参与竞选",
                    "type": "boolean"
                },
                "toMortgage": {
                    "description": "待抵押",
                    "type": "string"
                },
                "voted": {
                    "description": "选民贡献",
                    "type": "string"
                },
                "voters": {
                    "description": "投票人数",
                    "type": "integer"
                }
            }
        },
        "types.BeanVot": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "选民地址",
                    "type": "string"
                },
                "amount": {
                    "description": "投票金额",
                    "type": "string"
                }
            }
        },
...
        "types.BeanVots": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "节点公钥",
                    "type": "string"
                },
                "voted": {
                    "description": "选民贡献",
                    "type": "string"
                },
                "voters": {
                    "description": "选民列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.BeanVot"
                    }
                }
            }
        },
...
    }
}